Implement Degraded functionality.
This implements the Degraded functionality. Allowing tests to register a
Degraded status (which we've renamed from result).

To be in line with the "degraded" naming and due to the breaking change
of the TestFunc interface, we've renamed the Success and Failure verbs
to respectively "Available" and "Unavailable" which is more in line with
the "Degraded" verb. We've also renamed the "Result" interface name to
"Status" which again reflects these verbs better.