Make tests run concurrently.
This makes the health checks run concurrently. Allowing a quicker
execution speed.

This is useful for when a certain tests acts slower and could
potentially stagger other tests causing a slower overall response. If an
external probe has a timeout set, this could cause failures whilst in
practice it's only a single test that is slower, not failing.